Simplify fromOld functions
